requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
requests==2.18.4
urllib3==1.22
python-telegram-bot
psycopg2-binary
googlemaps
datetime